initial commit
This commit contains what I’ve worked on so far in Rails.
I’m using bootstrap to style all of the rails forms and pages, but I am
still in the process of doing that. This project is just a basic blog
application that I am building as I go through the Rails tutorial.